Shift PinnedOpportunityModel to user models
